% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{cofgeat}
\alias{cofgeat}
\title{User interface of cofgeat.}
\usage{
cofgeat(net, gexp, score = NULL)
}
\arguments{
\item{net}{gene regulatory network. format: a network of bnlearn is used.}

\item{gexp}{gene expression profile. format: a data frame containg the variables in the network.}

\item{score}{a character string, the label of the network score to be used in the algorithm. See bnlearn-package for details.}
}
\value{
dataframe of {genes} x {time points} that represent the controlling force of the genes at the time point.
}
\description{
User interface of cofgeat.
}

